Move update structure for test utils
