+++ src/Accordion/__tests__/AccordionHeader.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import createWithTheme from '../../../tests/utils/createWithTheme';
+import createWithTheme from '../../../utils/createWithTheme';
 import AccordionHeader from '../AccordionHeader';
 
 describe('<AccordionHeader />', () => {
