+++ src/Button/__tests__/Button.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import createWithTheme from '../../../tests/utils/createWithTheme';
+import createWithTheme from '../../../utils/createWithTheme';
 
 import Button from '../Button';
 import Icon from '../../Icon';
