+++ src/Checkbox/__tests__/Checkbox.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import createWithTheme from '../../../tests/utils/createWithTheme';
+import createWithTheme from '../../../utils/createWithTheme';
 import Checkbox from '../Checkbox';
 
 describe('<Checkbox />', () => {
