+++ src/Dropdown/__tests__/Dropdown.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import mountWithTheme from '../../../tests/utils/mountWithTheme';
+import mountWithTheme from '../../../utils/mountWithTheme';
 
 import Button from '../../Button';
 import Dropdown from '..';
