+++ src/Link/__tests__/Link.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import createWithTheme from '../../../tests/utils/createWithTheme';
+import createWithTheme from '../../../utils/createWithTheme';
 import Link from '../Link';
 
 describe('<Link />', () => {
