+++ utils/mountWithTheme.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { mount } from 'enzyme';
 import { ThemeProvider } from 'emotion-theming';
-import theme from '../../src/theme';
+import theme from '../src/theme';
 
 export default (Component, mountOptions = {}) => {
   const wrappedComponent = (
